- Fixed duplicated note when user invoke the map without saving - Fixed audio removed error when the audio is in memory but not in DB

+++ TaskAppsAssasinsIOS/Controller/Note/Repository/NoteDetailRepository.swift
@@ -20,7 +20,9 @@ extension NoteDetailViewController {
             request.predicate = predicate
             
             audiosEntity = try context.fetch(request)
-            context.delete(audiosEntity[0])
+            if audiosEntity.count > 0 {
+                context.delete(audiosEntity[0])
+            }
         }
         catch {
